+++ packages/client/src/Validator.ts
@@ -6,7 +6,6 @@ import {
     StreamMessage,
     StreamMessageValidator,
     SigningUtil,
-    StreamMessageError,
     StreamID,
     EthereumAddress
 } from 'streamr-client-protocol'
@@ -16,12 +15,6 @@ import { Stoppable } from './utils/Stoppable'
 import { Context } from './utils/Context'
 import { StreamRegistryCached } from './StreamRegistryCached'
 import { ConfigInjectionToken, SubscribeConfig, CacheConfig } from './Config'
-
-export class SignatureRequiredError extends StreamMessageError {
-    constructor(streamMessage: StreamMessage, code?: string) {
-        super('Client requires data to be signed.', streamMessage, code)
-    }
-}
 
 /**
  * Wrap StreamMessageValidator in a way that ensures it can validate in parallel but
@@ -75,12 +68,8 @@ export class Validator extends StreamMessageValidator implements Stoppable, Cont
         const { options } = this
 
         // Check special cases controlled by the verifySignatures policy
-        if (options.verifySignatures === 'never' && msg.messageType === StreamMessage.MESSAGE_TYPES.MESSAGE) {
+        if (options.verifySignatures === false && msg.messageType === StreamMessage.MESSAGE_TYPES.MESSAGE) {
             return // no validation required
-        }
-
-        if (options.verifySignatures === 'always' && !msg.signature) {
-            throw new SignatureRequiredError(msg)
         }
 
         // In all other cases validate using the validator